customer app core features completed
