+++ src/components/ThankYouOrder.js
@@ -1,13 +1,15 @@
 import React from 'react';
 import {Grid, Box, Typography, Paper, Button} from '@material-ui/core';
 import {makeStyles} from '@material-ui/core/styles';
+import useMediaQuery from '@material-ui/core/useMediaQuery';
 
 const ThankYouOrder = (props) => {
-    const searchMerchantName = props.searchMerchantName;
-    const matches = props.matches;
+    const matches = useMediaQuery('(min-width: 768px)');
+    //const searchMerchantName = props.searchMerchantName;
+    //const matches = props.matches;
     const order = props.order;
-    const handleClear = props.handleClear;
     const merchants = props.merchants;
+    const droneID = props.droneID;
 
     const classes = useStyles();
 
@@ -52,7 +54,12 @@ const ThankYouOrder = (props) => {
               </Grid>
           </Grid>
           <Box className={classes.clearButton}>
-            <Button size="small" className={classes.buttonClear} onClick={() => handleClear(order.id)}>
+            <Button 
+              size="small" 
+              className={classes.buttonClear} 
+              onClick={() => props.setState({...props.state,
+                show: props.state.show.set(order.id, false)
+              })}>
               Clear
             </Button>
           </Box>
@@ -103,5 +110,26 @@ const useStyles = makeStyles((theme) => ({
       background: '#E9E9E9',
     }
   }));
+
+  function searchMerchantName(id, arr) {
+    if (arr.length === 0) {
+        return null;
+    } else {
+        let len = arr.length;
+        let start = 0;
+        let end = len - 1;
+        
+        while (start < end) {
+            let mid = start + Math.floor((end - start)/2);
+            if (id <= arr[mid].id) {
+                end = mid;
+            } else {
+                start = mid + 1;
+            }
+        }
+
+        return (arr[start].id === id) ? arr[start].name : null;
+    }
+}
 
 export default ThankYouOrder;
